Saving system updated
The previous saving system was underwhelming, barely exporting the few
important matrices to a pickle, with a unique name. Now, the different
versions of the database used to create the exported hybrid_system are
used to create a folder in the folder Databases. The method to correct
double counting used and the endogenization of capitals (or not) are
present in the name of the folder. It allows to register multiple
versions of hybrid systems. The Leontief inverses are incorporated to
this folder as well.